requirements.txt
matplotlib==3.7.1
numpy==1.24.3
Pillow==11.1.0
scikit-image==0.22.0
torchmetrics==1.1.1
tqdm==4.65.0
